+++ dataflow/db/dbclients.py
@@ -3,7 +3,8 @@
 from influxdb_client import InfluxDBClient
 from influxdb_client import WriteOptions
 
-from modules import filereader
+# from modules import filereader
+# import modules.filereader as filereader
 
 
 def get_write_client(conf_db: dict):
@@ -14,8 +15,8 @@ def get_write_client(conf_db: dict):
     return client, write_client
 
 
-def get_query_client():
-    dbconf = filereader.read_configfile(config_file='../../../configs/dbconf.yaml')
-    client = InfluxDBClient(url=dbconf['url'], token=dbconf['token'], org=dbconf['org'])
-    query_client = client.query_api()
-    return client, query_client
+# def get_query_client():
+#     dbconf = filereader.read_configfile(config_file='../../../configs/dbconf.yaml')
+#     client = InfluxDBClient(url=dbconf['url'], token=dbconf['token'], org=dbconf['org'])
+#     query_client = client.query_api()
+#     return client, query_client